+++ src/commands/compose.ts
@@ -15,7 +15,8 @@
  */
 
 import { Command, flags } from '@oclif/command';
-import { BootstrapService, BootstrapUtils, CommandUtils, ComposeService } from '../service';
+import { BootstrapService, BootstrapUtils, ComposeService } from '../service';
+import { CommandUtils } from '../service/CommandUtils';
 
 export default class Compose extends Command {
     static description = 'It generates the `docker-compose.yml` file from the configured network.';